Refactored code to support relative/explicit imports
So, I thought I was clever having the API config data such as user_agent_client and user_agent_email be declared in
__init__.py, but it turns out that I was dumb, and that doing it that way requires that every module import the whole
library to access those values. When trying to change to relative imports as a best practice, I ran into issues where
suddenly these 'global' variables were not accessible (not to mention, ran into a whole lot of circular imports), so
I decided to scrap the whole idea and bring back the previously removed Weatherpy class as a sort of 'API base' class.
Now, users simply pass this object into the constructor for the various other classes, thus carrying forward their
configuration information. I also moved some of the API query logic into the Weatherpy class as it is going to be
needed for all of the endpoints, so I figured this was a good place to keep the code.